change id type to long

+++ src/main/java/com/maesproject/gtfs/repository/AlertRepository.java
@@ -8,7 +8,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 
 @Repository
-public interface AlertRepository extends JpaRepository<AlertEntity, Integer> {
+public interface AlertRepository extends JpaRepository<AlertEntity, Long> {
 
     List<AlertEntity> findByEntityId(String entityId);
